hw04 added test with spring boot test
