+++ HW04-spring-shell/src/test/resources/il8n/bundle_en.properties
@@ -1,16 +1,17 @@
 answer.correct=Answer accepted
 answer.incorrect=You have writen incorrect answer, follow the question instruction
-language.choose=Please choose the language using "l" and number of the langauage    
+exception.load=Failed to load questions
+language.choose=Please choose the language using "l" and number of the language
 language.show=Please use "ll" to show available languages
 locale.answer=print-answer:>
 locale.available=Available languages: {0}
 locale.choose=choose-locale:>
 locale.chosen=Chosen {0} language
-print.name=Print your name, uesing a command latine "a" or "answer"  
+print.name=Print your name, uesing a command latine "a" or "answer"
 print.offset.false=Уважаемый(ая) {0} {1} после ответа на {2} вопросов, у Вас НЕ ЗАЧЁТ
 print.offset.true=Уважаемый(ая) {0} {1} после ответа на {2} вопросов, у Вас ЗАЧЁТ
 print.result=Ваше результат: {0}
-print.surname=Print your surname, uesing a command latine "a" or "answer"  
+print.surname=Print your surname, useing a command latine "a" or "answer"
 question.multiple.text=Choose one correct options:\
 {0}\
 Options:\
