+++ HW04-spring-shell/src/test/resources/il8n/bundle_ru_RU.properties
@@ -1,5 +1,6 @@
 answer.correct=Ответ принят
 answer.incorrect=Вы ввели некорректный ответ, следуйте инструкциям к вопросу
+exception.load=Не удалось загрузить вопросы
 language.choose=Пожалуйста выберите язык, используя команду "l" и номер языка из списка
 language.show=Пожалуйста используйте "ll" чтобы посмотреть доступные языки
 locale.answer=введите-ответ:>
